requirements.txt
asgiref==3.2.10
beautifulsoup4==4.9.3
certifi==2020.12.5
cloudinary==1.25.0
confusable-homoglyphs==3.2.0
Django==3.1
django-bootstrap4==3.0.0
django-registration==3.1.2
django-tinymce==3.3.0
gunicorn==20.1.0
importlib-metadata==2.1.1

psycopg2==2.8.6
pytz==2021.1
six==1.15.0
soupsieve==2.2.1
sqlparse==0.4.1
urllib3==1.26.4
zipp==3.4.1
